rename sparse_map2d
