+++ source/consumers/Histogram2D.cpp
@@ -1,6 +1,6 @@
 #include <consumers/Histogram2D.h>
 
-#include <consumers/dataspaces/sparse_map2d.h>
+#include <consumers/dataspaces/SparseMap2D.h>
 #include <consumers/dataspaces/SparseMatrix2D.h>
 #include <consumers/dataspaces/dense_matrix2d.h>
