+++ tests/consumers/dataspaces/CMakeLists.txt
@@ -3,7 +3,7 @@ set(dir ${CMAKE_CURRENT_SOURCE_DIR})
 set(SOURCES
   ${dir}/dense1d.cpp
   ${dir}/scalar.cpp
-  ${dir}/sparse_map2d.cpp
+  ${dir}/SparseMap2DTest.cpp
   ${dir}/SparseMap3DTest.cpp
   ${dir}/SparseMatrix2DTest.cpp
   )
