+++ tests/consumers/dataspaces/SparseMap2DTest.cpp
@@ -1,6 +1,6 @@
 #include "gtest_color_print.h"
 
-#include <consumers/dataspaces/sparse_map2d.h>
+#include <consumers/dataspaces/SparseMap2D.h>
 
 class SparseMap2D : public TestBase
 {
